analysis: re-organize code/package structure
