+++ pkg/query/lang.go
@@ -1,0 +1,5 @@
+package query
+
+import "github.com/smacker/go-tree-sitter/python"
+
+var LanguagePython = python.GetLanguage()
