+++ pkg/query/location.go
@@ -1,0 +1,37 @@
+package query
+
+import (
+	sitter "github.com/smacker/go-tree-sitter"
+	"go.lsp.dev/protocol"
+
+	"github.com/tilt-dev/starlark-lsp/pkg/document"
+)
+
+// PositionToPoint converts an LSP protocol file location to a Tree-sitter file location.
+func PositionToPoint(pos protocol.Position) sitter.Point {
+	return sitter.Point{
+		Row:    pos.Line,
+		Column: pos.Character,
+	}
+}
+
+// PointToPosition converts a Tree-sitter file location to an LSP protocol file location.
+func PointToPosition(point sitter.Point) protocol.Position {
+	return protocol.Position{
+		Line:      point.Row,
+		Character: point.Column,
+	}
+}
+
+// NamedNodeAtPosition returns the most granular named descendant at a position.
+func NamedNodeAtPosition(doc document.Document, pos protocol.Position) (*sitter.Node, bool) {
+	pt := PositionToPoint(pos)
+	if doc.Tree == nil {
+		return nil, false
+	}
+	node := doc.Tree.RootNode().NamedDescendantForPointRange(pt, pt)
+	if node != nil {
+		return node, true
+	}
+	return nil, false
+}
