+++ pkg/query/node.go
@@ -1,0 +1,16 @@
+package query
+
+const (
+	NodeTypeArgList             = "argument_list"
+	NodeTypeFunctionDef         = "function_definition"
+	NodeTypeParameters          = "parameters"
+	NodeTypeIdentifier          = "identifier"
+	NodeTypeExpressionStatement = "expression_statement"
+	NodeTypeString              = "string"
+	NodeTypeBlock               = "block"
+
+	FieldName       = "name"
+	FieldParameters = "parameters"
+	FieldReturnType = "return_type"
+	FieldBody       = "body"
+)
